requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.3.0
backoff==2.2.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.15
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
effdet==0.4.1
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
filelock==3.16.1
filetype==1.2.0
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.1.0
google-api-core==2.24.0
google-auth==2.37.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.66.0
gotrue==2.9.2
greenlet==3.1.1
groq==0.15.0
grpcio==1.69.0
grpcio-status==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
keybert==0.8.5
kiwisolver==1.4.7
langchain==0.2.17
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langdetect==1.0.9
langsmith==0.1.147
layoutparser==0.3.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.15.1
opencv-python==4.11.0.86
orjson==3.10.15
packaging==24.2
pandas==2.0.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pg8000==1.31.2
phidata==2.7.9
pikepdf==9.2.1
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
portalocker==3.0.0
postgrest==0.16.11
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.1.4
pypdf==5.1.0
pypdfium2==4.30.1
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.10.22
python-magic==0.4.27
python-magic-bin==0.4.14
python-multipart==0.0.20
pytz==2024.2
pywin32==308
PyYAML==6.0.2
rapidfuzz==3.9.7
realtime==1.0.6
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
scramp==1.4.5
sentence-transformers==3.2.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
storage3==0.7.7
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
timm==1.0.14
tokenizers==0.20.3
tomli==2.2.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.46.3
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.11.8
unstructured-client==0.28.1
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.13
urllib3==2.2.3
websockets==12.0
wrapt==1.17.2
yarl==1.15.2
zipp==3.20.2
